custom callback to gather stat data
